+++ src/Registration/cReg/include/sirf/Reg/NiftyF3dSym.h
@@ -23,6 +23,10 @@ limitations under the License.
 \ingroup Registration
 \brief NiftyReg's f3d class for non-rigid registrations.
 
+The user has the choice to use the symmetric or non-symmetric version of the algorithm. 
+We believe there to be an upstream bug (https://github.com/KCL-BMEIS/niftyreg/issues/71)
+in the symmetric version, and therefore do not recommend using it until that issue is closed.
+
 \author Richard Brown
 \author CCP PETMR
 */
@@ -31,7 +35,7 @@ limitations under the License.
 
 #include "sirf/Reg/NiftyRegistration.h"
 
-template<class dataType> class reg_f3d_sym;
+template<class dataType> class reg_f3d;
 
 namespace sirf {
 
@@ -70,6 +74,10 @@ public:
     /// Set reference time point
     void set_reference_time_point(const int reference_time_point) { _reference_time_point = reference_time_point; }
 
+    /// Set use symmetric. Default is false. 
+    /// No need to expose to python/matlab until https://github.com/KCL-BMEIS/niftyreg/issues/71 is closed.
+    void set_use_symmetric(const bool use_symmetric) { _use_symmetric = use_symmetric; }
+
     /// Set initial affine transformation
     void set_initial_affine_transformation(const std::shared_ptr<const AffineTransformation<float> > mat) { _initial_transformation_sptr = mat; }
 
@@ -88,12 +96,14 @@ protected:
     void set_parameters();
 
     /// Registration object
-    std::shared_ptr<reg_f3d_sym<dataType> > _registration_sptr;
+    std::shared_ptr<reg_f3d<dataType> > _registration_sptr;
 
     /// Floating time point
     int _floating_time_point;
     /// Reference time point
     int _reference_time_point;
+    /// Use symmetric bool
+    bool _use_symmetric = false;
     /// Transformation matrix
     std::shared_ptr<const AffineTransformation<float> > _initial_transformation_sptr;
 };